Added Dropdown Component
Added Dropdown Component. Customized it so that it creates links using the Gatsby Link functionality. Also modified Dropdown to take two props -  dropdownName and dropdownLinks. 

dropdownLinks is an array of objectLiterals. Each objectLiteral has two values linkName and linkPath. linkName is what will be displayed. linkPath will be the destination.